all 3 fixtures
